apps/load-app/requirements.txt
pandas
boto3
python-dotenv
confluent_kafka